+++ nengo/utils/tests/test_graphs.py
@@ -1,19 +1,29 @@
+import collections
+
 from nengo.utils import graphs
 
 
+def graph(edges=None):
+    """Construct graph from edges."""
+    g = collections.defaultdict(set)
+    if edges is not None:
+        g.update(edges)
+    return g
+
+
 def test_reversedict():
-    edges = graphs.graph({"a": {"b", "c"}, "b": set(), "c": set()})
+    edges = graph({"a": {"b", "c"}, "b": set(), "c": set()})
     r_edges = graphs.reverse_edges(edges)
     assert r_edges == {"a": set(), "b": {"a"}, "c": {"a"}}
 
 
 def test_toposort():
-    edges = graphs.graph({"a": {"b", "c"}, "b": {"c"}, "c": set()})
+    edges = graph({"a": {"b", "c"}, "b": {"c"}, "c": set()})
     assert graphs.toposort(edges) == ["a", "b", "c"]
 
 
 def test_transitive_closure():
-    edges = graphs.graph({"a": {}, "b": {"c", "d"}, "c": set(), "d": {"e"}, "e": set()})
+    edges = graph({"a": {}, "b": {"c", "d"}, "c": set(), "d": {"e"}, "e": set()})
     assert graphs.transitive_closure(edges) == {
         "a": set(),
         "b": {"c", "d", "e"},
@@ -24,6 +34,6 @@ def test_transitive_closure():
 
 
 def test_add_edges():
-    edges = graphs.graph({"a": {"b", "c"}})
+    edges = graph({"a": {"b", "c"}})
     graphs.add_edges(edges, [("a", "d"), ("b", "c")])
     assert edges == {"a": {"b", "c", "d"}, "b": {"c"}}