+++ CHANGES.rst
@@ -22,7 +22,16 @@ Release history
 3.1.1 (unreleased)
 ==================
 
+**Fixed**
 
+- Operator graph step order will now be deterministic. (`#1654`_)
+
+**Removed**
+
+- Removed ``nengo.utils.graphs.graph`` (this was a small utility function for building
+  graphs that was only used in tests). (`#1654`_)
+
+.. _#1654: https://github.com/nengo/nengo/pull/1654
 
 3.1.0 (November 17, 2020)
 =========================